+++ components/profile/ImportedPublicationsSection.js
@@ -8,13 +8,16 @@ import api from '../../lib/api-client'
 import ProfileSection from './ProfileSection'
 
 const ImportedPublicationsSection = ({
+  profileId,
   updatePublicationIdsToUnlink,
-  publications,
-  totalCount,
+  reRender,
 }) => {
+  const { accessToken } = useUser()
+  const [publications, setPublications] = useState(null)
   const [publicationsToDisplay, setPublicationsToDisplay] = useState([])
   const [publicationIdsToUnlink, setPublicationIdsToUnlink] = useState([])
   const [pageNumber, setPageNumber] = useState(1)
+  const [totalCount, setTotalCount] = useState(0)
   const pageSize = 20
 
   const handleLinkUnlinkPublication = (id, isunlink = false) => {
@@ -36,16 +39,45 @@ const ImportedPublicationsSection = ({
     emptyMessage: 'No imported publications found',
   }
 
+  const loadPublications = async () => {
+    try {
+      const result = await api.getCombined(
+        '/notes',
+        {
+          'content.authorids': profileId,
+          invitations: ['dblp.org/-/record'],
+        },
+        {
+          'content.authorids': profileId,
+          invitations: [
+            'DBLP.org/-/Record',
+            'OpenReview.net/Archive/-/Imported_Record',
+            'OpenReview.net/Archive/-/Direct_Upload',
+          ],
+        },
+        { accessToken, includeVersion: true, sort: 'tmdate:desc' }
+      )
+      setPublications(result.notes)
+      setTotalCount(result.count)
+    } catch (error) {
+      promptError(`${error.message} when loading your publications`)
+    }
+  }
+
   useEffect(() => {
     updatePublicationIdsToUnlink(publicationIdsToUnlink)
   }, [publicationIdsToUnlink])
+
+  useEffect(() => {
+    loadPublications()
+  }, [reRender])
 
   useEffect(() => {
     if (!publications) return
     setPublicationsToDisplay(
       publications.slice((pageNumber - 1) * pageSize, pageNumber * pageSize)
     )
-  }, [pageNumber])
+  }, [pageNumber, publications])
 
   if (!publicationsToDisplay.length) {
     return (
